Updated filename and deleted irrelevant files
