+++ .gitignore
@@ -1,3 +1,4 @@
 *.egg-info
 .coverage
 __pycache__
+*.swp
