--- src/uwtools/resources/rocoto.xml
@@ -1,9 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE workflow [
-  <!ENTITY FOO "bar">
-  <!ENTITY BAZ "qux">
-]>
-<workflow>
-  <log>foo then &FOO;</log>
-  <task>baz then &BAZ;</task>
-</workflow>
