+++ src/uwtools/rocoto.py
@@ -1,51 +1,48 @@
-import xml.etree.ElementTree as ET
-from importlib import resources
-from pathlib import Path
-from typing import Any, Dict
+"""
+Support for writing a Rocoto XML using J2Templating
+"""
 
+from uwtools.j2template import J2Template
+from uwtools.config import YAMLConfig
 
-class RocotoXML:
+# private functions
+def _add_jobname(self, tasks: dict) -> None:
     """
-    A base class specifying methods to manipulate and generate XML files.
+    Add the jobname entry for all the tasks in the workflow.
+
+    :param tasks: Dict of tasks in workflow.
     """
+    pass
+    if not isinstance(tasks, dict):
+        return
+    for task, task_settings in tasks.items():
+        task_type = task.split("_", maxsplit=1)[0]
+        if task_type == "task":
+            # Use the provided attribute if it is present, otherwise use
+            # the name in the key
+            tasks[task]["jobname"] = \
+                task_settings.get("attrs", {}).get("name") or \
+                task.split("_", maxsplit=1)[1]
+        elif task_type == "metatask":
+            add_jobname(task_settings)
 
-    # Helper functions
 
-    def doctype(self, entities: Dict[str, Any]) -> str:
-        """
-        Generate the DOCTYPE section.
-        """
-        pass
+# Main method
+def write_rocoto_xml(self, input_yaml: str, input_template: str, rendered_output: str) -> None:
+    """
+    Main entry point.
 
-    def entity(self, name: str) -> str:
-        """
-        Return a reference to the given ENTITY name.
-        """
-        pass
+    :param input_yaml: Yaml file provided by the user.
+    :param input_template: Path to input template.
+    :param rendered_output: Path to directory to write  rendered XML file.
+    """
+    pass
+    # make an instance of the YAMLConfig class
+    # pass the input yaml to return a dict
+    config_obj = _load.YAMLConfig(input_yaml)
 
-    def fixup(self, xml: str, entities: Dict[str, str]) -> str:
-        """
-        Correct unwanted &-escaping.
-        """
-        pass
+    _add_jobname(config_obj)
 
-    # Main Methods
-
-    def __init__(self):
-        """
-        Construct XML document into a tree structure.
-        """
-        with resources.as_file(resources.files("uwtools.resources")) as path:
-            with open(path / "rocoto.xml", "r", encoding="utf-8") as f:
-                self._tree = ET.parse(f)
-
-    def write(self, output_dir: str) -> None:
-        """
-        Write rendered template to the directory provided.
-
-        :param output_path: Path to directory to write.
-        """
-        # pylint: disable=line-too-long
-        self._tree.write(
-            Path(output_dir) / "contents.xml", encoding="utf-8", xml_declaration=True, method="xml"
-        )
+   # render the template
+   #template = render_template.J2Template()
+
